Complete logic rework & code refactor
The fundamental change is that when computing an
operator, instead of relying on deducing the
correct action (What numbers to manipulate, how
to change the display, what should happens when a
number button is pressed...), it now works by
using an enum to change the overall "state" of the
calculator, which makes the intended behavior
explicit. Along the way several bugs were
encountered and fixed, and the code was overall
refactored.